backend/requirements.txt
# Sahayak Backend Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
Flask==2.3.0
Flask-CORS==4.0.0

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.0
lxml==4.9.0

# OCR and Image Processing
pytesseract==0.3.10
Pillow==11.0.0
pdf2image==1.16.3
PyPDF2==3.0.1

# Machine Learning
scikit-learn==1.3.0
numpy==1.24.0

# Utilities
python-dateutil==2.8.2
python-docx==1.1.0

# Optional: For better performance
# gunicorn==21.2.0  # For production deployment

# Note: You also need to install system dependencies:
# - Tesseract OCR: https://github.com/tesseract-ocr/tesseract
#   Ubuntu/Debian: sudo apt-get install tesseract-ocr
#   Windows: Download installer from GitHub releases
#   Mac: brew install tesseract
#
# - Poppler (for PDF processing):
#   Ubuntu/Debian: sudo apt-get install poppler-utils
#   Windows: Download from https://blog.alivate.com.au/poppler-windows/
#   Mac: brew install poppler